+++ scripts/promptops/__init__.py
@@ -1,0 +1,5 @@
+"""PromptOps tooling for ReleaseCopilot."""
+
+__all__ = [
+    "human_actions",
+]